+++ .gitignore
@@ -9,7 +9,7 @@ npm-debug.*
 *.mobileprovision
 *.orig.*
 web-build/
-.env
+
 # macOS
 .DS_Store
