+++ app/movie/[movieId].jsx
@@ -1,21 +1,11 @@
-import React, { useState, useEffect } from "react";
-import {
-  View,
-  Text,
-  Image,
-  TouchableOpacity,
-  StyleSheet,
-  Animated,
-  Pressable,
-  StatusBar,
-  ActivityIndicator,
-} from "react-native";
-import { AntDesign } from "@expo/vector-icons";
-import { Themes } from "../../constants/Themes";
-import { useRouter, useLocalSearchParams } from "expo-router";
-import RatingFavorite from "../../components/RatingFavorite";
-import { Ionicons } from "@expo/vector-icons";
-import { getMovieDetails, getMovieDirector } from "../../helpers/tmdbApi";
+import React, { useState, useEffect } from 'react';
+import { View, Text, Image, TouchableOpacity, StyleSheet, Animated, Pressable, StatusBar, ActivityIndicator } from 'react-native';
+import { AntDesign } from '@expo/vector-icons';
+import { Themes } from '../../constants/Themes';
+import { useRouter, useLocalSearchParams } from 'expo-router';
+import RatingFavorite from '../../components/RatingFavorite';
+import { Ionicons } from '@expo/vector-icons';
+import { getMovieDetails, getMovieDirector } from '../../helpers/tmdbApi';
 
 const MovieScreen = () => {
   const [showFullSynopsis, setShowFullSynopsis] = useState(false);
@@ -23,12 +13,13 @@ const MovieScreen = () => {
   const [measuredHeight, setMeasuredHeight] = useState(0);
   const [synopsisHeight, setSynopsisHeight] = useState(52);
   const scrollY = new Animated.Value(0);
-  const route = useRouter();
+  const route = useRouter()
   const { movieId } = useLocalSearchParams(); // Uso de useLocalSearchParams para obtener movieId
   const [movie, setMovie] = useState(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
-  const [director, setDirector] = useState("");
+  const [director, setDirector] = useState('');
+
   // Fetching detalles de la película
   useEffect(() => {
     const fetchMovieDetails = async () => {
@@ -46,88 +37,75 @@ const MovieScreen = () => {
     };
     fetchMovieDetails();
   }, [movieId]);
-
   useEffect(() => {
     const fetchDirector = async () => {
       try {
         const directorName = await getMovieDirector(movieId); // Usa la nueva función
         setDirector(directorName);
       } catch (error) {
-        console.error("Error fetching director:", error);
+        console.error('Error fetching director:', error);
       }
     };
-
+  
     fetchDirector();
   }, [movieId]);
 
+
   // Verificación de errores o datos vacíos
   if (loading) {
-    return (
-      <ActivityIndicator size="large" color="#6200EE" style={styles.loader} />
-    );
+    return <ActivityIndicator size="large" color="#6200EE" style={styles.loader} />;
   }
-
+  
   if (error) {
     return (
       <View style={styles.errorContainer}>
         <Text style={styles.errorText}>{error}</Text>
-        <TouchableOpacity
-          onPress={() => route.back()}
-          style={styles.backButtonError}
-        >
+        <TouchableOpacity onPress={() => route.back()} style={styles.backButtonError}>
           <AntDesign name="arrowleft" size={24} color="white" />
         </TouchableOpacity>
       </View>
     );
   }
-
+  
   // Condición adicional para asegurar que `movie` tiene datos válidos
   if (!movie || Object.keys(movie).length === 0) {
     return (
       <View style={styles.errorContainer}>
         <Text style={styles.errorText}>Movie details not available.</Text>
-        <TouchableOpacity
-          onPress={() => route.back()}
-          style={styles.backButtonError}
-        >
+        <TouchableOpacity onPress={() => route.back()} style={styles.backButtonError}>
           <AntDesign name="arrowleft" size={24} color="white" />
         </TouchableOpacity>
       </View>
     );
   }
-
   const toggleSynopsis = () => {
     setShowFullSynopsis(!showFullSynopsis);
     setSynopsisHeight(!showFullSynopsis ? measuredHeight : 52); // Expandir o contraer
   };
 
   const posterUri = movie.poster_path
-    ? `https://image.tmdb.org/t/p/w500${movie.poster_path}`
-    : "https://via.placeholder.com/500x750.png?text=No+Image";
+  ? `https://image.tmdb.org/t/p/w500${movie.poster_path}`
+  : 'https://via.placeholder.com/500x750.png?text=No+Image';
 
   const backdropUri = movie.backdrop_path
-    ? `https://image.tmdb.org/t/p/original${movie.backdrop_path}`
-    : "https://via.placeholder.com/500x750.png?text=No+Image";
+  ? `https://image.tmdb.org/t/p/original${movie.backdrop_path}`
+  : 'https://via.placeholder.com/500x750.png?text=No+Image';
 
   let rating = movie.vote_average / 2;
   rating = rating.toFixed(1);
 
   return (
     <View style={styles.container}>
-      <StatusBar
-        barStyle="light-content"
-        backgroundColor={Themes.colors.grayLight}
-      />
+      <StatusBar barStyle="light-content" backgroundColor={Themes.colors.grayLight} />
       <TouchableOpacity style={styles.backButton} onPress={() => route.back()}>
         <AntDesign name="arrowleft" size={24} color="white" />
       </TouchableOpacity>
 
       <Animated.ScrollView
-        onScroll={Animated.event(
+          onScroll={Animated.event(
           [{ nativeEvent: { contentOffset: { y: scrollY } } }],
           { useNativeDriver: true }
-        )}
-      >
+        )}>
         {/* Imagen de la película como header */}
         <Animated.View style={styles.headerImageContainer}>
           <Animated.Image
@@ -138,271 +116,220 @@ const MovieScreen = () => {
                 opacity: scrollY.interpolate({
                   inputRange: [0, 200],
                   outputRange: [1, 0],
-                  extrapolate: "clamp",
-                }),
-              },
+                  extrapolate: 'clamp'
+                })
+              }
             ]}
           />
         </Animated.View>
 
         {/* Información de la película */}
         <View style={styles.movieInfoContainer}>
-          <View style={styles.movieDetailsRow}>
+        <View style={styles.movieDetailsRow}>
             <View style={styles.movieTextContainer}>
               <Text style={styles.movieTitle}>{movie.title}</Text>
-              <Text style={styles.movieDirector}>Directed by</Text>
-              <Text style={styles.bold}>{director}</Text>
-              <Text style={styles.movieDetails}>
-                {movie.release_date} • {movie.runtime} min
+              <Text style={styles.movieDirector}>
+                Directed by 
               </Text>
+              <Text style={styles.bold}>{director}</Text>
+              <Text style={styles.movieDetails}>{movie.release_date} • {movie.runtime} min</Text>
             </View>
-            <Image source={{ uri: posterUri }} style={styles.movieThumbnail} />
+            <Image
+              source={{ uri: posterUri }}
+              style={styles.movieThumbnail}
+            />
           </View>
-          {/* Sinopsis */}
-          <View>
-            {/* Vista fantasma para medir la altura completa */}
-            <Text
-              style={[
-                styles.synopsisText,
-                { position: "absolute", opacity: 0, zIndex: -1 },
-              ]} // Oculta este texto
-              onLayout={(event) => {
-                const { height } = event.nativeEvent.layout;
-                if (height > 52) {
-                  setIsExpandable(true); // Activa el botón si es necesario
-                  setMeasuredHeight(height); // Guarda la altura completa
-                }
-              }}
-            >
+
+        {/* Sinopsis */}
+        <View>
+        {/* Vista fantasma para medir la altura completa */}
+          <Text
+            style={[styles.synopsisText, { position: 'absolute', opacity: 0, zIndex: -1 }]} // Oculta este texto
+            onLayout={(event) => {
+              const { height } = event.nativeEvent.layout;
+              if (height > 52) {
+                setIsExpandable(true); // Activa el botón si es necesario
+                setMeasuredHeight(height); // Guarda la altura completa
+              }
+            }}
+          >
+            {movie.overview || "No synopsis available."}
+          </Text>
+
+        {/* Texto visible con restricciones */}
+          <Animated.View style={{ height: showFullSynopsis ? measuredHeight : synopsisHeight }}>
+            <Text style={styles.synopsisText} numberOfLines={showFullSynopsis ? undefined : 3}>
               {movie.overview || "No synopsis available."}
             </Text>
+          </Animated.View>
 
-            {/* Texto visible con restricciones */}
-            <Animated.View
-              style={{
-                height: showFullSynopsis ? measuredHeight : synopsisHeight,
-              }}
-            >
-              <Text
-                style={styles.synopsisText}
-                numberOfLines={showFullSynopsis ? undefined : 3}
-              >
-                {movie.overview || "No synopsis available."}
+        {/* Botón de expansión */}
+          {isExpandable && (
+            <TouchableOpacity onPress={toggleSynopsis} style={styles.ex}>
+              <Text style={styles.moreText}>
+                {showFullSynopsis ? "Show Less" : "Show More"}
               </Text>
-            </Animated.View>
+            </TouchableOpacity>
+          )}
+        </View>
 
-            {/* Botón de expansión */}
-            {isExpandable && (
-              <TouchableOpacity onPress={toggleSynopsis} style={styles.ex}>
-                <Text style={styles.moreText}>
-                  {showFullSynopsis ? "Show Less" : "Show More"}
-                </Text>
-              </TouchableOpacity>
-            )}
-          </View>
           <View style={styles.divider} />
+
           {/* Rating */}
           <View style={styles.overallContainer}>
-            <View style={styles.ratingContainer}>
-              <Text style={styles.ratingLabel}>Overall Rating</Text>
-              <Text style={styles.ratingScore}>{rating}</Text>
 
-              <View>
-                <RatingFavorite
-                  style={styles.starsContainer}
-                  rating={rating}
-                  showFavorite={false}
-                  starSize={32} // Ajusta el tamaño si es necesario
-                />
-              </View>
+          <View style={styles.ratingContainer}>
+            <Text style={styles.ratingLabel}>Overall Rating</Text>
+            <Text style={styles.ratingScore}>{rating}</Text>
+
+            <View >
+            <RatingFavorite style={styles.starsContainer}
+                rating={rating}
+                showFavorite={false}
+                starSize={32}    // Ajusta el tamaño si es necesario
+            />
             </View>
 
-            <View style={styles.rateButtonContainer}>
-              <TouchableOpacity
-                style={styles.rateButton}
-                onPress={() => route.push(`/${movieId}`)}
-              >
-                <Text style={styles.rateButtonText}>Add this movie</Text>
-              </TouchableOpacity>
+          </View>
+
+          <View style={styles.rateButtonContainer}>
+            <TouchableOpacity style={styles.rateButton} onPress={() => route.push('movieReview')} >
+              <Text style={styles.rateButtonText}>Add this movie</Text>
+            </TouchableOpacity>
+          </View>
+
+          </View>
+
+          <View style={styles.divider} />
+
+          {/* Posts de los usuarios */}
+          <Text style={styles.postsTitle}>Posts</Text>
+          <Pressable onPress={() => route.push('post')}>
+          <View style={styles.postContainer}>
+          <Pressable style={styles.test} onPress={() => route.push('other_user_information')} >
+
+            <Ionicons name="person-circle-outline" size={48} color="#6200EE" style={styles.userPic} />
+
+          </Pressable>
+            <View style={styles.postContent}>
+              <View style={styles.usernameContainer}>
+                <Text style={styles.postUser}>User’s post</Text>
+                <View style={styles.userRating}>
+                  <RatingFavorite
+                    rating={5}
+                    isFavorite={true}
+                    starSize={16}    // Ajusta el tamaño si es necesario
+                    iconSize={12}    // Ajusta el tamaño si es necesario
+                  />
+                </View>
+              </View>
+              <Text style={styles.postText}>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum auctor magna ipsum, et egestas magna dapibus ac.</Text>
             </View>
           </View>
-          <View style={styles.divider} />|{/* Posts de los usuarios */}
-          <Text style={styles.postsTitle}>Posts</Text>
-          <Pressable onPress={() => route.push("post")}>
-            <View style={styles.postContainer}>
-              <Pressable
-                style={styles.test}
-                onPress={() => route.push("other_user_information")}
-              >
-                <Ionicons
-                  name="person-circle-outline"
-                  size={48}
-                  color="#6200EE"
-                  style={styles.userPic}
-                />
-              </Pressable>
-              <View style={styles.postContent}>
-                <View style={styles.usernameContainer}>
-                  <Text style={styles.postUser}>User’s post</Text>
-                  <View style={styles.userRating}>
-                    <RatingFavorite
-                      rating={5}
-                      isFavorite={true}
-                      starSize={16} // Ajusta el tamaño si es necesario
-                      iconSize={12} // Ajusta el tamaño si es necesario
-                    />
-                  </View>
-                </View>
-                <Text style={styles.postText}>
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  Vestibulum auctor magna ipsum, et egestas magna dapibus ac.
-                </Text>
-              </View>
-            </View>
           </Pressable>
           {/* Agrega más posts según sea necesario */}
-          <Pressable onPress={() => route.push("post")}>
-            <View style={styles.postContainer}>
-              <Pressable
-                style={styles.test}
-                onPress={() => route.push("other_user_information")}
-              >
-                <Ionicons
-                  name="person-circle-outline"
-                  size={48}
-                  color="#6200EE"
-                  style={styles.userPic}
-                />
-              </Pressable>
-              <View style={styles.postContent}>
-                <View style={styles.usernameContainer}>
-                  <Text style={styles.postUser}>User’s post</Text>
-                  <View style={styles.userRating}>
-                    <RatingFavorite
-                      rating={4.5}
-                      isFavorite={true}
-                      starSize={16} // Ajusta el tamaño si es necesario
-                      iconSize={12} // Ajusta el tamaño si es necesario
-                    />
-                  </View>
-                </View>
-                <Text style={styles.postText}>
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  Vestibulum auctor magna ipsum, et egestas magna dapibus ac.
-                </Text>
-              </View>
-            </View>
+          <Pressable onPress={() => route.push('post')}>
+          <View style={styles.postContainer}>
+          <Pressable style={styles.test} onPress={() => route.push('other_user_information')} >
+
+            <Ionicons name="person-circle-outline" size={48} color="#6200EE" style={styles.userPic} />
+
           </Pressable>
-          <Pressable onPress={() => route.push("post")}>
-            <View style={styles.postContainer}>
-              <Pressable
-                style={styles.test}
-                onPress={() => route.push("other_user_information")}
-              >
-                <Ionicons
-                  name="person-circle-outline"
-                  size={48}
-                  color="#6200EE"
-                  style={styles.userPic}
-                />
-              </Pressable>
-              <View style={styles.postContent}>
-                <View style={styles.usernameContainer}>
-                  <Text style={styles.postUser}>User’s post</Text>
-                  <View style={styles.userRating}>
-                    <RatingFavorite
-                      rating={4.3}
-                      isFavorite={false}
-                      starSize={16} // Ajusta el tamaño si es necesario
-                      iconSize={12} // Ajusta el tamaño si es necesario
-                    />
-                  </View>
+            <View style={styles.postContent}>
+              <View style={styles.usernameContainer}>
+                <Text style={styles.postUser}>User’s post</Text>
+                <View style={styles.userRating}>
+                  <RatingFavorite
+                    rating={4.5}
+                    isFavorite={true}
+                    starSize={16}    // Ajusta el tamaño si es necesario
+                    iconSize={12}    // Ajusta el tamaño si es necesario
+                  />
                 </View>
-                <Text style={styles.postText}>
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  Vestibulum auctor magna ipsum, et egestas magna dapibus ac.
-                </Text>
               </View>
+              <Text style={styles.postText}>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum auctor magna ipsum, et egestas magna dapibus ac.</Text>
             </View>
+          </View>
           </Pressable>
-          <Pressable onPress={() => route.push("post")}>
-            <View style={styles.postContainer}>
-              <Pressable
-                style={styles.test}
-                onPress={() => route.push("other_user_information")}
-              >
-                <Ionicons
-                  name="person-circle-outline"
-                  size={48}
-                  color="#6200EE"
-                  style={styles.userPic}
-                />
-              </Pressable>
-              <View style={styles.postContent}>
-                <View style={styles.usernameContainer}>
-                  <Text style={styles.postUser}>User’s post</Text>
-                  <View style={styles.userRating}>
-                    <RatingFavorite
-                      rating={4.4}
-                      isFavorite={true}
-                      starSize={16} // Ajusta el tamaño si es necesario
-                      iconSize={12} // Ajusta el tamaño si es necesario
-                    />
-                  </View>
-                </View>
-                <Text style={styles.postText}>
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  Vestibulum auctor magna ipsum, et egestas magna dapibus ac.
-                </Text>
-              </View>
-            </View>
+          <Pressable onPress={() => route.push('post')}>
+          <View style={styles.postContainer}>
+          <Pressable style={styles.test} onPress={() => route.push('other_user_information')} >
+
+            <Ionicons name="person-circle-outline" size={48} color="#6200EE" style={styles.userPic} />
+
           </Pressable>
-          <Pressable onPress={() => route.push("post")}>
-            <View style={styles.postContainer}>
-              <Pressable
-                style={styles.test}
-                onPress={() => route.push("other_user_information")}
-              >
-                <Ionicons
-                  name="person-circle-outline"
-                  size={48}
-                  color="#6200EE"
-                  style={styles.userPic}
-                />
-              </Pressable>
-              <View style={styles.postContent}>
-                <View style={styles.usernameContainer}>
-                  <Text style={styles.postUser}>User’s post</Text>
-                  <View style={styles.userRating}>
-                    <RatingFavorite
-                      rating={4.6}
-                      isFavorite={true}
-                      starSize={16} // Ajusta el tamaño si es necesario
-                      iconSize={12} // Ajusta el tamaño si es necesario
-                    />
-                  </View>
+            <View style={styles.postContent}>
+              <View style={styles.usernameContainer}>
+                <Text style={styles.postUser}>User’s post</Text>
+                <View style={styles.userRating}>
+                  <RatingFavorite
+                    rating={4.3}
+                    isFavorite={false}
+                    starSize={16}    // Ajusta el tamaño si es necesario
+                    iconSize={12}    // Ajusta el tamaño si es necesario
+                  />
                 </View>
-                <Text style={styles.postText}>
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  Vestibulum auctor magna ipsum, et egestas magna dapibus ac.
-                </Text>
               </View>
+              <Text style={styles.postText}>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum auctor magna ipsum, et egestas magna dapibus ac.</Text>
             </View>
+          </View>
+          </Pressable>
+          <Pressable onPress={() => route.push('post')}>
+          <View style={styles.postContainer}>
+          <Pressable style={styles.test} onPress={() => route.push('other_user_information')} >
+
+            <Ionicons name="person-circle-outline" size={48} color="#6200EE" style={styles.userPic} />
+
+          </Pressable>
+            <View style={styles.postContent}>
+              <View style={styles.usernameContainer}>
+                <Text style={styles.postUser}>User’s post</Text>
+                <View style={styles.userRating}>
+                  <RatingFavorite
+                    rating={4.4}
+                    isFavorite={true}
+                    starSize={16}    // Ajusta el tamaño si es necesario
+                    iconSize={12}    // Ajusta el tamaño si es necesario
+                  />
+                </View>
+              </View>
+              <Text style={styles.postText}>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum auctor magna ipsum, et egestas magna dapibus ac.</Text>
+            </View>
+          </View>
+          </Pressable>
+          <Pressable onPress={() => route.push('post')}>
+          <View style={styles.postContainer}>
+          <Pressable style={styles.test} onPress={() => route.push('other_user_information')} >
+
+            <Ionicons name="person-circle-outline" size={48} color="#6200EE" style={styles.userPic} />
+
+          </Pressable>
+            <View style={styles.postContent}>
+              <View style={styles.usernameContainer}>
+                <Text style={styles.postUser}>User’s post</Text>
+                <View style={styles.userRating}>
+                  <RatingFavorite
+                    rating={4.6}
+                    isFavorite={true}
+                    starSize={16}    // Ajusta el tamaño si es necesario
+                    iconSize={12}    // Ajusta el tamaño si es necesario
+                  />
+                </View>
+              </View>
+              <Text style={styles.postText}>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum auctor magna ipsum, et egestas magna dapibus ac.</Text>
+            </View>
+          </View>
           </Pressable>
         </View>
       </Animated.ScrollView>
 
       {/* Botón flotante */}
-      <TouchableOpacity
-        style={styles.floatingButton}
-        onPress={() => route.push(`/${movieId}`)}
-      >
+      <TouchableOpacity style={styles.floatingButton} onPress={() => route.push('movieReview')}>
         <AntDesign name="plus" size={24} color="white" />
       </TouchableOpacity>
     </View>
   );
 };
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -410,22 +337,22 @@ const styles = StyleSheet.create({
   },
   headerImageContainer: {
     height: 170,
-    position: "relative",
+    position: 'relative',
     backgroundColor: Themes.colors.grayLight,
   },
   test: {
-    alignSelf: "flex-start",
+    alignSelf: 'flex-start',
   },
   headerImage: {
-    width: "100%",
-    height: "100%",
+    width: '100%',
+    height: '100%',
     opacity: 0.9, // Opacidad inicial para efecto de difuminado
   },
   backButton: {
-    position: "absolute",
+    position: 'absolute',
     top: 10,
     left: 10,
-    backgroundColor: "rgba(0,0,0,0.5)",
+    backgroundColor: 'rgba(0,0,0,0.5)',
     borderRadius: 50,
     padding: 5,
     zIndex: 1000,
@@ -436,107 +363,107 @@ const styles = StyleSheet.create({
   movieTitle: {
     marginTop: -16,
     fontSize: 24,
-    color: "white",
-    fontWeight: "bold",
+    color: 'white',
+    fontWeight: 'bold',
   },
   movieDirector: {
-    color: "white",
+    color: 'white',
     fontSize: 16,
     marginTop: 8,
   },
   bold: {
-    fontWeight: "bold",
-    color: "white",
+    fontWeight: 'bold',
+    color: 'white',
     fontSize: 16,
-    marginBottom: 8 + 4,
+    marginBottom: 8+4,
   },
   movieDetails: {
-    color: "white",
+    color: 'white',
     fontSize: 14,
     marginBottom: 10,
   },
   synopsisContainer: {
-    flexDirection: "column",
-    alignItems: "flex-start", // Asegura alineación adecuada
-    overflow: "hidden", // Esconde texto fuera de la altura inicial
-    width: "100%", // O el tamaño que necesites
-  },
+    flexDirection: 'column',
+    alignItems: 'flex-start', // Asegura alineación adecuada
+    overflow: 'hidden', // Esconde texto fuera de la altura inicial
+    width: '100%', // O el tamaño que necesites
+  },  
   synopsisText: {
     marginTop: 8,
-    color: "white",
+    color: 'white',
     flex: 1,
   },
   moreDots: {
-    color: "#6200EE",
+    color: '#6200EE',
     fontSize: 20,
-    marginTop: -8 + 4,
-    marginBottom: -12 + 4,
+    marginTop: -8+4,
+    marginBottom: -12+4,
   },
   ratingContainer: {
-    flexDirection: "column",
-    alignItems: "center",
+    flexDirection: 'column',
+    alignItems: 'center',
   },
   ratingLabel: {
-    color: "white",
+    color: 'white',
     fontSize: 18,
   },
   ratingScore: {
-    color: "#FFF",
+    color: '#FFF',
     fontSize: 20,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
   starsContainer: {
-    flexDirection: "row",
+    flexDirection: 'row',
     zIndex: 1000,
   },
   rateButtonContainer: {
-    justifyContent: "center",
-    alignItems: "center",
+    justifyContent: 'center',
+    alignItems: 'center',
     flex: 1,
   },
   rateButton: {
-    backgroundColor: "#6200EE",
+    backgroundColor: '#6200EE',
     padding: 10,
     borderRadius: 5,
     marginLeft: 16,
   },
   rateButtonText: {
-    color: "white",
+    color: 'white',
     fontSize: 14,
   },
   postsTitle: {
-    color: "white",
+    color: 'white',
     fontSize: 18,
-    fontWeight: "bold",
+    fontWeight: 'bold',
     marginTop: 0,
   },
   postContainer: {
-    flexDirection: "row",
-    alignItems: "center",
+    flexDirection: 'row',
+    alignItems: 'center',
     paddingVertical: 10,
     borderBottomWidth: 1,
-    borderBottomColor: "#887c9f",
+    borderBottomColor: '#887c9f',
   },
   postContent: {
     flex: 1,
     paddingLeft: 10,
   },
   postUser: {
-    color: "white",
+    color: 'white',
     fontSize: 16,
   },
   userRating: {
     marginLeft: 8,
   },
   postText: {
-    color: "white",
+    color: 'white',
     fontSize: 14,
   },
   floatingButton: {
-    position: "absolute",
+    position: 'absolute',
     bottom: 20,
     right: 20,
-    backgroundColor: "#6200EE",
+    backgroundColor: '#6200EE',
     borderRadius: 50,
     padding: 15,
   },
@@ -550,21 +477,21 @@ const styles = StyleSheet.create({
     marginRight: 6,
   },
   overallContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
   },
   usernameContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
+    flexDirection: 'row',
+    justifyContent: 'space-between',
   },
   userPic: {
     marginTop: 0,
   },
   movieDetailsRow: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
   },
   movieThumbnail: {
     width: 80,
@@ -576,12 +503,13 @@ const styles = StyleSheet.create({
   },
   synopsisText: {
     fontSize: 14,
-    color: "white",
+    color: 'white',
   },
   ex: {
     flex: 1,
-    alignItems: "center",
-    justifyContent: "space-between",
+    alignItems: 'center',
+    justifyContent: 'space-between',
+
   },
   movieTextContainer: {
     flex: 0.95,
@@ -589,7 +517,7 @@ const styles = StyleSheet.create({
   moreText: {
     color: Themes.colors.purpleDetail,
     fontSize: 14,
-    alignItems: "center",
+    alignItems: 'center',
   },
 });
